continue implement translatable attributes

+++ lib/dynamic/schema/klass.rb
@@ -57,19 +57,17 @@ module Dynamic
           validates_presence_of :const_table_name
 
           after_create :create_table
-          after_create :create_translation_table
-
           before_destroy :rename_table_for_destruction
-          before_destroy :rename_translation_table_for_destruction
+
+          include Translation
         end
 
-        def drop_table # for maintenance
+        def drop_tables(do_it = false) # for maintenance
+          $stderr.puts 'dry run:' unless do_it
           return unless const_table_name.present?
+          $stderr.puts "drop #{const_table_name}"
+          return unless do_it
           self.class.connection.drop_table(const_table_name)
-        end
-
-        def const_translation_table_name
-          "#{const_table_name}_translations"
         end
 
         private
@@ -108,27 +106,6 @@ module Dynamic
           end
         end
 
-        def create_translation_table
-          return unless const_table_name.present? && !self.class.connection.data_sources.include?(const_translation_table_name)
-
-          self.class.connection.create_table(const_translation_table_name) do |t|
-            t.integer :record_id
-            t.index :record_id, name: "index_d_translations_#{self.class.base_class.name.gsub(/\//, '_').underscore}_#{self.id}"
-            Attribute::Translatable::Base::PRIMITIVE_SUBCLASS_NAMES.each do |n|
-              k = n.constantize
-              for i in 0..(k::MAX_INDEXED_COLUMN - 1)
-                t.send(k.column_type, :"#{k.column_name_prefix}i#{i}", index: true)
-              end
-              for i in 0..(k::MAX_NOT_INDEXED_COLUMN - 1)
-                t.send(k.column_type, :"#{k.column_name_prefix}#{i}", index: false)
-              end
-            end
-            t.timestamps
-            t.string :locale # indexed ?
-          end
-
-        end
-
         def rename_table_for_destruction
           return false unless const_table_name.present?
           self.original_const_table_name = self.const_table_name
@@ -139,14 +116,52 @@ module Dynamic
           return result
         end
 
-        def rename_translation_table_for_destruction
-          return self.class.connection.rename_table("#{original_const_table_name}_translations", const_translation_table_name)
-        end
-
         module Translation; extend ActiveSupport::Concern
 
+          included do
+            after_create :create_translation_table
+            before_destroy :rename_translation_table_for_destruction
+          end
+
+          def const_translation_table_name
+            "#{const_table_name}_translations"
+          end
+
+          def drop_tables(do_it = false) # for maintenance
+            super
+            return unless const_translation_table_name.present?
+            $stderr.puts "drop #{const_translation_table_name}"
+            return unless do_it
+            self.class.connection.drop_table(const_translation_table_name)
+          end
+
+          private
+
+          def create_translation_table
+            return unless const_table_name.present? && !self.class.connection.data_sources.include?(const_translation_table_name)
+
+            self.class.connection.create_table(const_translation_table_name) do |t|
+              t.integer :record_id
+              t.index :record_id, name: "index_d_translations_#{self.class.base_class.name.gsub(/\//, '_').underscore}_#{self.id}"
+              Attribute::Translatable::PRIMITIVE_SUBCLASS_NAMES.each do |n|
+                k = n.constantize
+                for i in 0..(k::MAX_INDEXED_COLUMN - 1)
+                  t.send(k.column_type, :"#{k.column_name_prefix}i#{i}", index: true)
+                end
+                for i in 0..(k::MAX_NOT_INDEXED_COLUMN - 1)
+                  t.send(k.column_type, :"#{k.column_name_prefix}#{i}", index: false)
+                end
+              end
+              t.timestamps
+              t.string :locale # indexed ?
+            end
+
+          end
+
+          def rename_translation_table_for_destruction
+            return self.class.connection.rename_table("#{original_const_table_name}_translations", const_translation_table_name)
+          end
         end
-        include Translation
 
       end
       include DatabaseManagement
